Split up classes from viewer_interaction.py
